Swift: reorganize schema lib files
